+++ src/main/java/hellfirepvp/astralsorcery/common/perk/tree/PerkTreeConstellation.java
@@ -119,7 +119,7 @@ public class PerkTreeConstellation<T extends AbstractPerk> extends PerkTreePoint
 
         double uLength = tex.getULength();
         double vLength = tex.getVLength();
-        Tuple<Double, Double> frameUV = tex.getUVOffset(spriteOffsetTick);
+        Tuple<Float, Float> frameUV = tex.getUVOffset(spriteOffsetTick);
 
         for (int i = 0; i < 4; i++) {
             int u = ((i + 1) & 2) >> 1;
